Move test utils functions to utils
